Checkpoint before pulling in latest queue code.
