+++ airbyte-integrations/bases/bases-destination-jdbc/src/main/java/io/airbyte/integrations/destination/staging/StagingConsumerFactory.java
@@ -290,85 +290,11 @@ public class StagingConsumerFactory {
     return streamDescToWriteConfig;
   }
 
-  class StagingAsyncFlusher implements StreamDestinationFlusher {
-
-    private final Map<StreamDescriptor, WriteConfig> streamDescToWriteConfig;
-    private final StagingOperations stagingOperations;
-    private final JdbcDatabase database;
-    private final ConfiguredAirbyteCatalog catalog;
-
-    public StagingAsyncFlusher(final Map<StreamDescriptor, WriteConfig> streamDescToWriteConfig,
-                               final StagingOperations stagingOperations,
-                               final JdbcDatabase database,
-                               final ConfiguredAirbyteCatalog catalog) {
-      this.streamDescToWriteConfig = streamDescToWriteConfig;
-      this.stagingOperations = stagingOperations;
-      this.database = database;
-      this.catalog = catalog;
-    }
-
-    // todo(davin): exceptions are too broad.
-    @Override
-    public void flush(final StreamDescriptor decs, final Stream<AirbyteMessage> stream) throws Exception {
-      // write this to a file - serilizable buffer?
-      // where do we create all the write configs?
-      CsvSerializedBuffer writer = null;
-      try {
-        writer = new CsvSerializedBuffer(
-            new FileBuffer(CsvSerializedBuffer.CSV_GZ_SUFFIX),
-            new StagingDatabaseCsvSheetGenerator(),
-            true);
-
-        CsvSerializedBuffer finalWriter = writer;
-        stream.forEach(record -> {
-          try {
-            // todo(davin): handle non-record airbyte messages.
-            finalWriter.accept(record.getRecord());
-          } catch (Exception e) {
-            throw new RuntimeException(e);
-          }
-        });
-      } catch (Exception e) {
-        throw new RuntimeException(e);
-      }
-
-      writer.flush();
-      LOGGER.info("Flushing buffer for stream {} ({}) to staging", decs.getName(), FileUtils.byteCountToDisplaySize(writer.getByteCount()));
-      if (!streamDescToWriteConfig.containsKey(decs)) {
-        throw new IllegalArgumentException(
-            String.format("Message contained record from a stream that was not in the catalog. \ncatalog: %s", Jsons.serialize(catalog)));
-      }
-
-      final WriteConfig writeConfig = streamDescToWriteConfig.get(decs);
-      final String schemaName = writeConfig.getOutputSchemaName();
-      final String stageName = stagingOperations.getStageName(schemaName, writeConfig.getStreamName());
-      final String stagingPath =
-          stagingOperations.getStagingPath(RANDOM_CONNECTION_ID, schemaName, writeConfig.getStreamName(), writeConfig.getWriteDatetime());
-      try {
-        LOGGER.info("Starting upload to stage..");
-        final String stagedFile = stagingOperations.uploadRecordsToStage(database, writer, schemaName, stageName, stagingPath);
-        copyIntoTableFromStage(database, stageName, stagingPath, List.of(stagedFile), writeConfig.getOutputTableName(), schemaName,
-            stagingOperations);
-      } catch (final Exception e) {
-        LOGGER.error("Failed to flush and commit buffer data into destination's raw table", e);
-        throw new RuntimeException("Failed to upload buffer to stage and commit to destination", e);
-      }
-
-      writer.close();
-    }
-
-    @Override
-    public long getOptimalBatchSizeBytes() {
-      return 200 * 1024 * 1024;
-    }
-
-  }
-
   /**
    * Handles copying data from staging area to destination table and clean up of staged files if
    * upload was unsuccessful
    */
-  private void copyIntoTableFromStage(final JdbcDatabase database,
+  public static void copyIntoTableFromStage(final JdbcDatabase database,
                                       final String stageName,
                                       final String stagingPath,
                                       final List<String> stagedFiles,
